+++ src/plugins/vis_type_timeline/public/to_ast.test.ts
@@ -38,7 +38,7 @@ describe('timeline vis toExpressionAst function', () => {
   beforeEach(() => {
     vis = {
       params: {
-        expression: '.es(*)',
+        expression: '.opensearch(*)',
         interval: 'auto',
       },
     } as any;
@@ -50,7 +50,7 @@ describe('timeline vis toExpressionAst function', () => {
   });
 
   it('should not escape single quotes', () => {
-    vis.params.expression = `.es(index=my*,timefield="date",split='test field:3',metric='avg:value')`;
+    vis.params.expression = `.opensearch(index=my*,timefield="date",split='test field:3',metric='avg:value')`;
     const actual = toExpressionAst(vis);
     expect(actual).toMatchSnapshot();
   });